fix: move utils to own file to avoid non-node-16 deps
